pyproject.toml
[project]
name = "dweller-avatar"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "minio>=7.2.15",
    "openai>=1.63.2",
    "pydantic-settings>=2.7.1",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.1",
]

[dependency-groups]
dev = [
    "ruff>=0.9.6",
]
